+++ src/c_console.c
@@ -39,7 +39,6 @@
 #include <Windows.h>
 #endif
 
-#include "SDL_image.h"
 #include "SDL_mixer.h"
 
 #include "am_map.h"
@@ -4044,8 +4043,6 @@ void C_PrintSDLVersions(void)
     C_Output(ITALICS("%s") " is using v%i.%i.%i of the " ITALICS("SDL (Simple DirectMedia Layer)") " library.",
         DOOMRETRO_NAME, SDL_MAJOR_VERSION, SDL_MINOR_VERSION, SDL_PATCHLEVEL);
 
-    C_Output("It is also using v%i.%i.%i of the " ITALICS("SDL_mixer")
-        " library and v%i.%i.%i of the " ITALICS("SDL_image") " library.",
-        SDL_MIXER_MAJOR_VERSION, SDL_MIXER_MINOR_VERSION, SDL_MIXER_PATCHLEVEL,
-        SDL_IMAGE_MAJOR_VERSION, SDL_IMAGE_MINOR_VERSION, SDL_IMAGE_PATCHLEVEL);
+    C_Output("It is also using v%i.%i.%i of the " ITALICS("SDL_mixer") " library.",
+        SDL_MIXER_MAJOR_VERSION, SDL_MIXER_MINOR_VERSION, SDL_MIXER_PATCHLEVEL);
 }